+++ src/main/js/types/PullRequest.ts
@@ -37,6 +37,8 @@ export type BasicPullRequest = {
 export type PullRequest = BasicPullRequest & HalRepresentation & {
   description?: string;
   author: DisplayedUser;
+  reviser?: DisplayedUser;
+  closeDate?: string;
   id: string;
   creationDate: string;
   reviewer: Reviewer[];